requirements.txt
streamlit==1.28.0
langgraph==0.0.20
langchain-groq==0.1.0
langchain-core==0.1.30
python-dotenv==1.0.0
pymongo==4.5.0